+++ libvast/src/system/writer_command.cpp
@@ -11,40 +11,28 @@
  * contained in the LICENSE file.                                             *
  ******************************************************************************/
 
-#pragma once
+#include "vast/system/writer_command.hpp"
 
-#include "vast/fwd.hpp"
-#include "vast/system/accountant.hpp"
-#include "vast/system/instrumentation.hpp"
+#include "vast/logger.hpp"
+#include "vast/system/make_sink.hpp"
+#include "vast/system/sink_command.hpp"
 
-#include <caf/behavior.hpp>
-#include <caf/fwd.hpp>
-#include <caf/stateful_actor.hpp>
+#include <caf/actor.hpp>
+#include <caf/make_message.hpp>
 
-#include <chrono>
-#include <cstdint>
+#include <string>
 
 namespace vast::system {
 
-// The base class for SINK actors.
-struct sink_state {
-  std::chrono::steady_clock::duration flush_interval = std::chrono::seconds(1);
-  std::chrono::steady_clock::time_point last_flush;
-  uint64_t processed = 0;
-  uint64_t max_events = 0;
-  caf::event_based_actor* self;
-  caf::actor statistics_subscriber;
-  accountant_type accountant;
-  vast::system::measurement measurement;
-  format::writer_ptr writer;
-  const char* name = "writer";
-
-  explicit sink_state(caf::event_based_actor* self_ptr);
-
-  void send_report();
-};
-
-caf::behavior sink(caf::stateful_actor<sink_state>* self,
-                   format::writer_ptr&& writer, uint64_t max_events);
+command::fun make_writer_command(std::string_view format) {
+  return [format = std::string{format}](const invocation& inv,
+                                        caf::actor_system& sys) {
+    VAST_TRACE(inv);
+    auto snk = make_sink(sys, format, inv.options);
+    if (!snk)
+      return make_message(snk.error());
+    return caf::make_message(sink_command(inv, sys, std::move(*snk)));
+  };
+}
 
 } // namespace vast::system