+++ libvast/vast/system/get_command.hpp
@@ -13,38 +13,13 @@
 
 #pragma once
 
-#include "vast/fwd.hpp"
-#include "vast/system/accountant.hpp"
-#include "vast/system/instrumentation.hpp"
+#include "vast/command.hpp"
 
-#include <caf/behavior.hpp>
 #include <caf/fwd.hpp>
-#include <caf/stateful_actor.hpp>
-
-#include <chrono>
-#include <cstdint>
 
 namespace vast::system {
 
-// The base class for SINK actors.
-struct sink_state {
-  std::chrono::steady_clock::duration flush_interval = std::chrono::seconds(1);
-  std::chrono::steady_clock::time_point last_flush;
-  uint64_t processed = 0;
-  uint64_t max_events = 0;
-  caf::event_based_actor* self;
-  caf::actor statistics_subscriber;
-  accountant_type accountant;
-  vast::system::measurement measurement;
-  format::writer_ptr writer;
-  const char* name = "writer";
-
-  explicit sink_state(caf::event_based_actor* self_ptr);
-
-  void send_report();
-};
-
-caf::behavior sink(caf::stateful_actor<sink_state>* self,
-                   format::writer_ptr&& writer, uint64_t max_events);
+/// Gets events by ID.
+caf::message get_command(const invocation& invocation, caf::actor_system& sys);
 
 } // namespace vast::system